requirements.txt
scikit-learn==0.22.2.post1
numpy==1.18.2
pandas==1.0.3
python-Levenshtein==0.12.0